requirements.txt
streamlit>=1.35.0
ultralytics>=8.2.0
opencv-python>=4.9.0.80
pillow>=10.3.0
numpy>=1.26.0
